Revert "Swap VDD calculation"
This was based on the incorrect assumption that:

vdd_now / vdd_at_cal = vref_now / vref_at_cal

vref is *not* a voltage divider but a bandgap.  So when vdd decreases, the ADC
full range decreases with it and vref increases. . So the right formula is what
we had before:

vdd_now / vdd_at_cal = vref_at_cal / vref_now

Reference manual RM0376, section 14.9 has the correct formula with full
explanation.

This reverts commit 60ea2729998fc08a9185e86bfd3aae036ffec25a.